Rename platform to system
