+++ internal/system/cmd/module/operatemodule/switchmodule_test.go
